[day15] Implement part2
I will admit this runs... very slowly. But it does work.

I'm certain my quite-special-casey shunting of both parts of the box could be nicer, and the "backout and roll back" behaviour is not the nicest (we're copying the entire map for every move even if it's not possible to hit the problematic scenario), but hey, the code works for the use case we have. And at least I can understand what it's doing.