[day19] Implement parts 1 and 2
I would normally commit these separately, but I was on a good run (pretty
chuffed with the 4 minute delta).

My original implementation already had the memo, but was using it to map to
booleans (is this solvable?) rather than integers.

The inner loop also returned as soon as a match was found, rather than
completing the full iteration.

Modifying that was pretty straightforward, just needing a new wrapper function
to get the sum (as my previous one just returned an array of patterns thinking
we might change those somehow...)