[day20] Only consider nodes reachable once
This speeds up everything as we're not accumulating a growing set at each step nor are we considering them all for cheats.

I imagine the next faster thing is finding a quicker way of computing all the reachable sets (this isn't actually that fast on my machine), but it's fast enough.